Flask/requirements.txt
streamlit==0.69.2
pandas
numpy 
sklearn
ics
flask 
newsapi-python
itsdangerous
requests
pytz
datetime
